Move RichTextPlugin to plugin folder
